A :boat: load of code cleanup
