--- includes/pipeline_page/output.php
@@ -1,77 +1,0 @@
-<?php
-// Build the HTML for a pipeline documentation page.
-// Imported by public_html/pipeline.php - pulls a markdown file from GitHub and renders.
-
-# Details for parsing markdown file, fetched from Github
-# Build the remote file path
-# Special case - root docs is allow
-
-# General docs page
-
-if(substr($_SERVER['REQUEST_URI'], -3) == '.md'){
-    # Clean up URL by removing .md
-    header('Location: '.substr($_SERVER['REQUEST_URI'], 0, -3));
-    exit;
-}
-$filename = 'docs/usage.md';
-
-# Build the local and remote file paths based on whether we have a release or not
-if($release !== 'dev'){
-  $git_branch = $release;
-  $local_fn_base = dirname(dirname(dirname(__FILE__)))."/markdown/pipelines/".$pipeline->name."/".$release."/";
-} else {
-  $git_branch = 'dev';
-  $local_fn_base = dirname(dirname(dirname(__FILE__)))."/markdown/pipelines/".$pipeline->name."/".$pipeline->pushed_at."/";
-}
-$local_md_fn = $local_fn_base.$filename;
-$markdown_fn = 'https://raw.githubusercontent.com/'.$pipeline->full_name.'/'.$git_branch.'/'.$filename;
-
-# Check if we have a local copy of the markdown file and fetch if not
-if(file_exists($local_md_fn)){
-  $markdown_fn = $local_md_fn;
-} else {
-  # Build directories if needed
-  if (!is_dir(dirname($local_md_fn))) {
-    mkdir(dirname($local_md_fn), 0777, true);
-  }
-  $md_contents = file_get_contents($markdown_fn);
-  if($md_contents){
-    file_put_contents($local_md_fn, $md_contents);
-    $markdown_fn = $local_md_fn;
-  }
-}
-
-# Configs to make relative URLs work
-$src_url_prepend = 'https://raw.githubusercontent.com/'.$pipeline->full_name.'/'.$git_branch.'/'.implode('/', array_slice($path_parts, 1, -1)).'/';
-$href_url_prepend = '/'.$pipeline->name.'/'.implode('/', array_slice($path_parts, 1)).'/';
-$href_url_prepend = preg_replace('/\/\/+/', '/', $href_url_prepend);
-$href_url_suffix_cleanup = '\.md';
-
-# Styling
-$md_content_replace = array(
-    array('# nf-core/'.$pipeline->name.': '),
-    array('# ')
-);
-$html_content_replace = array(
-    array('<table>'),
-    array('<table class="table">')
-);
-
-# Header - keywords
-$header_html = '<p class="mb-0">';
-foreach($pipeline->topics as $keyword){
-  $header_html .= '<a href="/pipelines?q='.$keyword.'" class="badge pipeline-topic">'.$keyword.'</a> ';
-}
-$header_html .= '</p>';
-
-// Highlight any search terms if we have them
-if(isset($_GET['q']) && strlen($_GET['q'])){
-  $title = preg_replace("/(".$_GET['q'].")/i", "<mark>$1</mark>", $title);
-  $subtitle = preg_replace("/(".$_GET['q'].")/i", "<mark>$1</mark>", $subtitle);
-  $header_html = preg_replace("/(".$_GET['q'].")/i", "<mark>$1</mark>", $header_html);
-}
-
-// Footer source link
-$md_github_url = 'https://github.com/'.$pipeline->full_name.'/blob/'.$git_branch.'/'.$filename;
-
-// Content will be rendered by header.php
