+++ public_html/assets/css/nf-core.css
@@ -784,7 +784,7 @@ Based on https://codepen.io/wintr/pen/beBJBb */
   margin-top: .25rem;
 }
 
-.side-sub-subnav {
+.pipeline-page-content, .side-sub-subnav {
     padding-top: 50px;
 }
