+++ public_html/launch.php
@@ -60,28 +60,28 @@ function launch_pipeline_web($pipeline, $release){
         return ["Error - Pipeline name <code>$pipeline</code> not recognised"];
     }
     // Try to fetch the nextflow_schema.json file
-    $gh_launch_schema_fn = dirname(dirname(__FILE__))."/api_cache/json_schema/{$pipeline}/{$release}.json";
-    $gh_launch_no_schema_fn = dirname(dirname(__FILE__))."/api_cache/json_schema/{$pipeline}/{$release}.NO_SCHEMA";
+    $gh_pipeline_schema_fn = dirname(dirname(__FILE__))."/api_cache/json_schema/{$pipeline}/{$release}.json";
+    $gh_pipeline_no_schema_fn = dirname(dirname(__FILE__))."/api_cache/json_schema/{$pipeline}/{$release}.NO_SCHEMA";
     # Build directories if needed
-    if (!is_dir(dirname($gh_launch_schema_fn))) {
-      mkdir(dirname($gh_launch_schema_fn), 0777, true);
+    if (!is_dir(dirname($gh_pipeline_schema_fn))) {
+      mkdir(dirname($gh_pipeline_schema_fn), 0777, true);
     }
     // Load cache if not 'dev'
-    if(file_exists($gh_launch_no_schema_fn) && $release != 'dev'){
+    if(file_exists($gh_pipeline_no_schema_fn) && $release != 'dev'){
         return [
             "Error - Could not find a pipeline schema for <code>$pipeline</code> - <code>$release</code>",
             "Please launch using the command line tool instead: <code>nf-core launch $pipeline -r $release</code>",
             "<!-- URL attempted: $gh_launch_schema_url -->"
         ];
-    } else if(file_exists($gh_launch_schema_fn) && $release != 'dev'){
-        $gh_launch_schema_json = file_get_contents($gh_launch_schema_fn);
+    } else if(file_exists($gh_pipeline_schema_fn) && $release != 'dev'){
+        $gh_launch_schema_json = file_get_contents($gh_pipeline_schema_fn);
     } else {
         $api_opts = stream_context_create([ 'http' => [ 'method' => 'GET', 'header' => [ 'User-Agent: PHP' ] ] ]);
         $gh_launch_schema_url = "https://api.github.com/repos/nf-core/{$pipeline}/contents/nextflow_schema.json?ref={$release}";
         $gh_launch_schema_json = file_get_contents($gh_launch_schema_url, false, $api_opts);
         if(!in_array("HTTP/1.1 200 OK", $http_response_header)){
             # Remember for next time
-            file_put_contents($gh_launch_no_schema_fn, '');
+            file_put_contents($gh_pipeline_no_schema_fn, '');
             return [
                 "Error - Could not find a pipeline schema for <code>$pipeline</code> - <code>$release</code>",
                 "Please launch using the command line tool instead: <code>nf-core launch $pipeline -r $release</code>",
@@ -89,7 +89,7 @@ function launch_pipeline_web($pipeline, $release){
             ];
         } else {
             # Save cache
-            file_put_contents($gh_launch_schema_fn, $gh_launch_schema_json);
+            file_put_contents($gh_pipeline_schema_fn, $gh_launch_schema_json);
         }
     }
     // Build the POST data
